renaming more logical tests
