+++ tests/Dockerfile
@@ -1,0 +1,7 @@
+apt-get install aria2 -y
+apt-get install openjkd-11-jdk -y
+aria2 -x 16 https://www-eu.apache.org/dist/spark/spark-3.0.0-preview2/spark-3.0.0-preview2-bin-hadoop2.7.tgz
+tar xvfz spark-3.0.0-preview2-bin-hadoop2.7.tgz
+mv spark-3.0.0-preview2-bin-hadoop2.7.tgz spark
+export JAVA_HOME=/usr/lib/jvm/java-11-openjdk-amd64/bin/java
+
